homework2/src/requirements.txt
Flask
minio
flask-jwt-extended
python-keycloak